* commit 'ab3d57e3326ff98052e07f2cd22b59d6536bf96c':
  Improved truncate for mysql and mongodb

+++ src/test/java/com/knubisoft/testlum/testing/framework/db/mongodb/MongoOperation.java
@@ -37,7 +37,9 @@ public class MongoOperation implements StorageOperation {
         mongoDatabases.forEach((aliasEnv, database) -> {
             if (Objects.equals(aliasEnv.getEnvironment(), EnvManager.currentEnv())) {
                 for (String collectionName : database.listCollectionNames()) {
-                    database.getCollection(collectionName).drop();
+                    if (database.getCollection(collectionName).countDocuments() > 0) {
+                        database.getCollection(collectionName).drop();
+                    }
                 }
             }
         });
